pyproject.toml
[project]
name = "distribution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dacite>=1.9.2",
    "gmsh>=4.13.1",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.6",
    "meshio>=5.3.5",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "pyvista[all]>=0.43.0",
    "tqdm>=4.67.1",
    "trame>=3.10.2",
]
